requirements.txt
Django==4.1.5
psycopg2==2.9.5
python-dotenv==0.21.1
django-leaflet==0.28.3
djangorestframework==3.14.0
